+++ lib/unified_esc_pos_printer.dart
@@ -14,6 +14,7 @@ export 'src/core/enums.dart';
 export 'src/core/extensions.dart';
 export 'src/core/generator.dart';
 export 'src/core/print_column.dart';
+export 'src/core/print_raster_column.dart';
 export 'src/core/qrcode.dart';
 export 'src/core/print_text_styles.dart';
 export 'src/core/ticket.dart';